re: refactor to proper directory structure
